Revert "soc: intel_adsp: rework host IPC using IPC service"
This reverts commit cf7e2e63c17aab88289637ebc691003de1c8b628.

This commit introduced a regression in Sound Open Firmware for LNL and
PTL platforms. Tests executing DSP power flows do not receive FW Ready
notifications on the second iteration. From preliminary debugging, it
appears that the DSP does not enter interrupt after sending the
notification.
