+++ soc/intel/intel_adsp/common/CMakeLists.txt
@@ -9,7 +9,7 @@ zephyr_library_named(intel_adsp_common)
 zephyr_include_directories(include)
 zephyr_library_include_directories(${ZEPHYR_BASE}/drivers)
 
-zephyr_library_sources_ifdef(CONFIG_INTEL_ADSP_IPC_OLD_INTERFACE ipc.c)
+zephyr_library_sources_ifdef(CONFIG_INTEL_ADSP_IPC ipc.c)
 
 zephyr_library_sources_ifdef(CONFIG_GDBSTUB
     gdbstub_backend_sram.c
