+++ subsys/ipc/ipc_service/backends/Kconfig
@@ -50,5 +50,4 @@ config IPC_SERVICE_BACKEND_ICMSG_ME_FOLLOWER
 
 rsource "Kconfig.icmsg_me"
 rsource "Kconfig.icbmsg"
-rsource "Kconfig.intel_adsp"
 rsource "Kconfig.rpmsg"
